fix: more improvements in table column menu
