+++ packages/@tinacms/fields/src/Wysiwyg/state/plugins/Menu/FloatingTableMenu/TableOptionsMenu.tsx
@@ -19,45 +19,14 @@ limitations under the License.
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { EditorView } from 'prosemirror-view'
-import { deleteTable, TableMap } from 'prosemirror-tables'
+import { deleteTable } from 'prosemirror-tables'
 import styled from 'styled-components'
-import {
-  findParentNodeOfType,
-  forEachCellInColumn,
-  setCellAttrs,
-} from 'prosemirror-utils'
 
-import { AlignCenter, AlignLeft, AlignRight, TrashIcon } from '@tinacms/icons'
+import { TrashIcon } from '@tinacms/icons'
 import { IconButton } from '@tinacms/styles'
 
 interface TableOptionsMenu {
   view: EditorView
-}
-
-const alignColumn = (view: EditorView, alignValue: string) => {
-  const { state, dispatch } = view
-  const { selection } = state
-  const { table, table_cell, table_header } = state.schema.nodes
-  const tableCell = findParentNodeOfType(table_cell)(state.selection)
-  const tableHeader = findParentNodeOfType(table_header)(state.selection)
-  const cellNode = tableCell || tableHeader
-  if (!cellNode) return
-  const align = cellNode.node.attrs.align === alignValue ? undefined : alignValue
-  const tableNode = findParentNodeOfType(table)(state.selection)
-  if (!tableNode) return
-  const tableMap = TableMap.get(tableNode.node)
-  const pos = Object.entries(tableMap.map).find(
-    entry => entry[1] > selection.head - tableNode.start
-  )
-  if (!pos) return
-  const cellPos = parseInt(pos[0]) - 1
-  const columnPos = cellPos % tableMap.width
-  dispatch(
-    forEachCellInColumn(columnPos, (cell, tr) =>
-      setCellAttrs(cell, { align })(tr)
-    )(state.tr)
-  )
-  view.focus()
 }
 
 export default ({ view }: TableOptionsMenu) => {
@@ -75,15 +44,6 @@ export default ({ view }: TableOptionsMenu) => {
   const { height, width } = tableElm.getBoundingClientRect()
   return ReactDOM.createPortal(
     <Wrapper height={height} width={width}>
-      <IconButton onClick={() => alignColumn(view, 'left')} small primary>
-        <AlignLeft />
-      </IconButton>
-      <IconButton onClick={() => alignColumn(view, 'center')} small primary>
-        <AlignCenter />
-      </IconButton>
-      <IconButton onClick={() => alignColumn(view, 'right')} small primary>
-        <AlignRight />
-      </IconButton>
       <IconButton onClick={deleteSelectedTable} small primary>
         <TrashIcon />
       </IconButton>
@@ -97,13 +57,9 @@ const Wrapper = styled.div<
 >`
   background-color: #ffffff;
   border-radius: 2px;
-  display: flex;
   cursor: default;
   padding: 0px 4px;
   position: absolute;
   top: ${({ height }) => `${height + 24}px`};
-  left: ${({ width }) => `${width / 2 - 64}px`};
-  button:not(:first-of-type) {
-    margin-left: 10px;
-  }
+  left: ${({ width }) => `${width / 2 - 8}px`};
 `
